src/gradio/requirements.txt
requests
python-dotenv
gradio==4.14.0
